+++ src/evaluation_system/api/plugin.py
@@ -26,7 +26,7 @@ from __future__ import annotations
 import abc
 import logging
 import os
-import re
+import random
 import shlex
 import shutil
 import socket
@@ -39,15 +39,21 @@ from configparser import ConfigParser, ExtendedInterpolation
 from contextlib import contextmanager
 from datetime import datetime
 from functools import partial
+from hashlib import sha512
 from pathlib import Path
 from time import time
-from typing import IO, Any, Dict, Iterable, Iterator, Optional, TextIO, Union, cast
+from typing import (
+    IO,
+    Any,
+    Dict,
+    Iterable,
+    Iterator,
+    Optional,
+    TextIO,
+    Union,
+    cast,
+)
 from uuid import uuid4
-
-from PyPDF2 import PdfReader
-from rich.console import Console
-from rich.file_proxy import FileProxy
-from rich.traceback import Traceback
 
 import evaluation_system.model.history.models as hist_model
 import evaluation_system.model.repository as repository
@@ -61,6 +67,10 @@ from evaluation_system.misc.exceptions import (
 from evaluation_system.misc.utils import PIPE_OUT, TemplateDict
 from evaluation_system.model.solr_core import SolrCore
 from evaluation_system.model.user import User
+from PyPDF2 import PdfReader
+from rich.console import Console
+from rich.file_proxy import FileProxy
+from rich.traceback import Traceback
 
 from .user_data import DataReader
 from .workload_manager import JobStatus, schedule_job
@@ -347,7 +357,14 @@ class PluginAbstract(abc.ABC):
                 self._user.getUserSchedulerOutputDir(),
                 plugin_name.lower(),
             )
-            self._plugin_out = Path(log_directory) / f"{plugin_name}-{pid}.local"
+            # We need to make the output file unique, in order to do so
+            # we generate a more or less random sample from the current
+            # hostname and the system time
+            suffix = sha512((socket.gethostname() + str(time())).encode())
+            rand_suffix = "".join(random.choices(suffix.hexdigest(), k=8))
+            self._plugin_out = (
+                Path(log_directory) / f"{plugin_name}-{rand_suffix}-{pid}.local"
+            )
         return self._plugin_out
 
     def _set_interactive_job_as_running(self, rowid: Optional[int]):